#99 Implementing TestScheduler
- Add class TestMultiThreadScheduler
- Rename older Test of Scheduler and create new TestDiscreteTimeMultiThreadScheduler and TestRealTimeMultiThreadScheduler
- All test for scheduleOnce and schedulerExecutable are done
- Miss of test of schedule repeated, infinitely, scheduleEvent
