+++ src/test/java/sima/core/scheduler/TestScheduler.java
@@ -3,11 +3,11 @@ package sima.core.scheduler;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import sima.core.scheduler.multithread.DiscreteTimeMultiThreadScheduler;
-import sima.core.scheduler.multithread.RealTimeMultiThreadScheduler;
+import sima.core.scheduler.exception.NotSchedulableTimeException;
 
 import java.util.List;
 import java.util.Vector;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -15,11 +15,19 @@ import static org.junit.jupiter.api.Assertions.*;
  * Test that all Classes which implements {@link Scheduler} must pass.
  */
 @Disabled
-public class TestScheduler {
+public abstract class TestScheduler {
 
     // Static.
 
     protected static Scheduler SCHEDULER;
+
+    /**
+     * To specify and must be the same which is specify when {@link #SCHEDULER} is instantiate.
+     * <p>
+     * In that way we can verify if the Scheduler returns the right value with the method
+     * {@link Scheduler#getEndSimulation()}.
+     */
+    protected static long END_SIMULATION;
 
     /**
      * Define the tolerance when we test when the scheduler execute executable. Example: If a executable must be execute
@@ -28,7 +36,29 @@ public class TestScheduler {
      */
     protected static long TIME_EXECUTION_TOLERANCE = 0;
 
+    // Setup.
+
+    @BeforeEach
+    public void setup() {
+        this.initialize();
+    }
+
+    /**
+     * Here in this function you initialize all need variable for tests.
+     * <p>
+     * Needed variables are: <strong>{@link #END_SIMULATION}, {@link #SCHEDULER},
+     * {@link #TIME_EXECUTION_TOLERANCE}</strong>
+     */
+    protected abstract void initialize();
+
     // Tests.
+
+    @Test
+    public void endSimulationIsCorrect() {
+        assertEquals(END_SIMULATION, SCHEDULER.getEndSimulation());
+
+        assertTrue(END_SIMULATION >= 2);
+    }
 
     @Test
     public void canAddOneSchedulerWatcher() {
@@ -220,7 +250,9 @@ public class TestScheduler {
         assertTrue(SCHEDULER.start());
         assertFalse(SCHEDULER.isRunning());
 
+        assertEquals(1, watcher.isPassToSchedulerKilled);
         assertEquals(1, watcher.isPassToNoExecutionToExecute);
+        assertEquals(0, watcher.isPassToSimulationEndTimeReach);
     }
 
     @Test
@@ -314,7 +346,39 @@ public class TestScheduler {
 
         List<Executable> executables = new Vector<>();
 
-        TestExecutableFeeder eF0 = new TestExecutableFeeder(Scheduler.NOW, executables);
+        TestExecutableFeederV1 eF0 = new TestExecutableFeederV1(Scheduler.NOW, executables);
+
+        SCHEDULER.scheduleExecutable(eF0, Scheduler.NOW, Scheduler.ScheduleMode.ONCE, -1, -1);
+
+        assertTrue(SCHEDULER.start());
+
+        // Finish by not reaching time.
+        blockSchedulerWatcher.waitUntilKilled();
+
+        for (Executable executable : executables) {
+            TestExecutableFeederV1 executableFeeder = (TestExecutableFeederV1) executable;
+            if (executableFeeder.executedTime != -1)
+                // Executable executed
+                this.verifyNumber(executableFeeder.executedTime, executableFeeder.timeToBeExecuted,
+                        TIME_EXECUTION_TOLERANCE);
+            else
+                // Executable normally out of the end of the simulation
+                assertTrue(executableFeeder.timeToBeExecuted > SCHEDULER.getEndSimulation());
+        }
+    }
+
+    @Test
+    public void watcherNotifyOneTimeWhenSchedulerFinishByReachingEndSimulation() {
+        BlockSchedulerWatcher blockSchedulerWatcher = new BlockSchedulerWatcher();
+        SCHEDULER.addSchedulerWatcher(blockSchedulerWatcher);
+
+        TestSchedulerWatcher watcher = new TestSchedulerWatcher();
+        SCHEDULER.addSchedulerWatcher(watcher);
+
+        // Useless.
+        List<Executable> executables = new Vector<>();
+
+        TestExecutableFeederV1 eF0 = new TestExecutableFeederV1(Scheduler.NOW, executables);
 
         SCHEDULER.scheduleExecutable(eF0, Scheduler.NOW, Scheduler.ScheduleMode.ONCE, -1, -1);
 
@@ -322,8 +386,114 @@ public class TestScheduler {
 
         blockSchedulerWatcher.waitUntilKilled();
 
+        assertEquals(1, watcher.isPassToSchedulerKilled);
+        assertEquals(0, watcher.isPassToNoExecutionToExecute);
+        assertEquals(1, watcher.isPassToSimulationEndTimeReach);
+    }
+
+    @Test
+    public void scheduleAtSpecificTimeThrowsExceptionIfSpecifiedTimeIsLessThanOne() {
+        TestExecutable e0 = new TestExecutable();
+
+        assertThrows(IllegalArgumentException.class,
+                () -> SCHEDULER.scheduleExecutableAtSpecificTime(e0, 0));
+        assertThrows(IllegalArgumentException.class,
+                () -> SCHEDULER.scheduleExecutableAtSpecificTime(e0, -1));
+    }
+
+    @Test
+    public void scheduleAtSpecificTimeThrowsExceptionIfSpecifiedTimeIsAlreadyPass() {
+        BlockSchedulerWatcher blockSchedulerWatcher = new BlockSchedulerWatcher();
+        SCHEDULER.addSchedulerWatcher(blockSchedulerWatcher);
+
+        TestSchedulerWatcher watcher = new TestSchedulerWatcher();
+        SCHEDULER.addSchedulerWatcher(watcher);
+
+        long timeToBeExecuted = (long) (0.5 * SCHEDULER.getEndSimulation());
+
+        final AtomicBoolean isPassed = new AtomicBoolean(false);
+
+        Executable e = () -> {
+            isPassed.set(true);
+            this.verifyNumber(SCHEDULER.getCurrentTime(), timeToBeExecuted, TIME_EXECUTION_TOLERANCE);
+
+            // Try to schedule an Executable at a passed timed.
+            assertThrows(NotSchedulableTimeException.class,
+                    () -> SCHEDULER.scheduleExecutableAtSpecificTime(new TestExecutable(),
+                            SCHEDULER.getCurrentTime()));
+        };
+
+        try {
+            SCHEDULER.scheduleExecutable(e, timeToBeExecuted, Scheduler.ScheduleMode.ONCE, -1, -1);
+        } catch (Exception exc) {
+            fail();
+        }
+
+        assertTrue(SCHEDULER.start());
+
+        // Finish by not reaching time.
+        blockSchedulerWatcher.waitUntilKilled();
+
+        assertTrue(isPassed.get());
+
+        assertEquals(1, watcher.isPassToSchedulerKilled);
+        assertEquals(1, watcher.isPassToNoExecutionToExecute);
+        assertEquals(0, watcher.isPassToSimulationEndTimeReach);
+    }
+
+    @Test
+    public void scheduleAtSpecificTimeScheduleAndExecuteAnExecutableAtTheSpecifiedTime() {
+        BlockSchedulerWatcher blockSchedulerWatcher = new BlockSchedulerWatcher();
+        SCHEDULER.addSchedulerWatcher(blockSchedulerWatcher);
+
+        long timeToBeExecuted = (long) (0.75 * SCHEDULER.getEndSimulation());
+
+        final AtomicBoolean isPassed = new AtomicBoolean(false);
+        Executable e = () -> {
+            isPassed.set(true);
+            this.verifyNumber(SCHEDULER.getCurrentTime(), timeToBeExecuted, TIME_EXECUTION_TOLERANCE);
+        };
+
+        try {
+            SCHEDULER.scheduleExecutableAtSpecificTime(e, timeToBeExecuted);
+        } catch (Exception exc) {
+            fail(exc);
+        }
+
+        assertTrue(SCHEDULER.start());
+
+        // Finish by not reaching time.
+        blockSchedulerWatcher.waitUntilKilled();
+
+        assertTrue(isPassed.get());
+    }
+
+    @Test
+    public void scheduleOnceThrowsExceptionIfWaitingIsLessOrEqualToZero() {
+        assertThrows(IllegalArgumentException.class,
+                () -> SCHEDULER.scheduleExecutableOnce(new TestExecutable(), 0));
+        assertThrows(IllegalArgumentException.class,
+                () -> SCHEDULER.scheduleExecutableOnce(new TestExecutable(), -1));
+    }
+
+    @Test
+    public void scheduleOnceScheduleAndExecuteOneTimeAnExecutableAtTime() {
+        BlockSchedulerWatcher blockSchedulerWatcher = new BlockSchedulerWatcher();
+        SCHEDULER.addSchedulerWatcher(blockSchedulerWatcher);
+
+        List<Executable> executables = new Vector<>();
+
+        TestExecutableFeederV2 eF0 = new TestExecutableFeederV2(Scheduler.NOW, executables);
+
+        SCHEDULER.scheduleExecutableOnce(eF0, Scheduler.NOW);
+
+        assertTrue(SCHEDULER.start());
+
+        // Finish by not reaching time.
+        blockSchedulerWatcher.waitUntilKilled();
+
         for (Executable executable : executables) {
-            TestExecutableFeeder executableFeeder = (TestExecutableFeeder) executable;
+            TestExecutableFeederV2 executableFeeder = (TestExecutableFeederV2) executable;
             if (executableFeeder.executedTime != -1)
                 // Executable executed
                 this.verifyNumber(executableFeeder.executedTime, executableFeeder.timeToBeExecuted,
@@ -424,17 +594,17 @@ public class TestScheduler {
         }
     }
 
-    protected static class TestExecutableFeeder implements Executable {
+    protected static class TestExecutableFeederV1 implements Executable {
 
         // Variables.
 
-        private final List<Executable> executableList;
         private final long timeToBeExecuted;
+        private final List<Executable> executableList;
         private long executedTime = -1;
 
         // Constructors.
 
-        public TestExecutableFeeder(long timeToBeExecuted, List<Executable> executableList) {
+        public TestExecutableFeederV1(long timeToBeExecuted, List<Executable> executableList) {
             this.executableList = executableList;
             this.executableList.add(this);
 
@@ -445,9 +615,37 @@ public class TestScheduler {
 
         @Override
         public void execute() {
-            TestExecutableFeeder executableFeeder =
-                    new TestExecutableFeeder(SCHEDULER.getCurrentTime() + 10, this.executableList);
+            TestExecutableFeederV1 executableFeeder =
+                    new TestExecutableFeederV1(SCHEDULER.getCurrentTime() + 10, this.executableList);
             SCHEDULER.scheduleExecutable(executableFeeder, 10, Scheduler.ScheduleMode.ONCE, -1, -1);
+            this.executedTime = SCHEDULER.getCurrentTime();
+        }
+    }
+
+    protected static class TestExecutableFeederV2 implements Executable {
+
+        // Variables.
+
+        private final long timeToBeExecuted;
+        private final List<Executable> executableList;
+        private long executedTime = -1;
+
+        // Constructors.
+
+        public TestExecutableFeederV2(long timeToBeExecuted, List<Executable> executableList) {
+            this.executableList = executableList;
+            this.executableList.add(this);
+
+            this.timeToBeExecuted = timeToBeExecuted;
+        }
+
+        // Methods.
+
+        @Override
+        public void execute() {
+            TestExecutableFeederV2 executableFeederV2 =
+                    new TestExecutableFeederV2(SCHEDULER.getCurrentTime() + 10, this.executableList);
+            SCHEDULER.scheduleExecutableOnce(executableFeederV2, 10);
             this.executedTime = SCHEDULER.getCurrentTime();
         }
     }
