+++ src/test/java/sima/core/scheduler/multithread/TestMultiThreadScheduler.java
@@ -1,0 +1,13 @@
+package sima.core.scheduler.multithread;
+
+import org.junit.jupiter.api.Disabled;
+import sima.core.scheduler.TestScheduler;
+
+@Disabled
+public abstract class TestMultiThreadScheduler extends TestScheduler {
+
+    // Variables.
+
+    protected int NB_EXECUTOR_THREADS = 5;
+
+}
